+++ src/main/java/com/project/coalba/domain/schedule/service/BossScheduleService.java
@@ -1,9 +1,5 @@
 package com.project.coalba.domain.schedule.service;
 
-import com.project.coalba.domain.auth.entity.User;
-import com.project.coalba.domain.externalCalendar.dto.CalendarEventDto;
-import com.project.coalba.domain.externalCalendar.dto.CalendarPersonalDto;
-import com.project.coalba.domain.externalCalendar.service.ExternalCalendarService;
 import com.project.coalba.domain.profile.entity.Staff;
 import com.project.coalba.domain.profile.service.StaffProfileService;
 import com.project.coalba.domain.schedule.entity.enums.*;
@@ -29,7 +25,6 @@ public class BossScheduleService {
     private final BossWorkspaceService bossWorkspaceService;
     private final StaffProfileService staffProfileService;
     private final ScheduleRepository scheduleRepository;
-    private final ExternalCalendarService externalCalendarService;
 
     @Transactional(readOnly = true)
     public BossHomePageServiceDto getHomePage() {
@@ -103,27 +98,23 @@ public class BossScheduleService {
     }
 
     @Transactional
-    public void save(ScheduleCreateServiceDto serviceDto) {
+    public Schedule save(ScheduleCreateServiceDto serviceDto) {
         Workspace workspace = bossWorkspaceService.getWorkspace(serviceDto.getWorkspaceId());
         Staff staff = staffProfileService.getStaff(serviceDto.getStaffId());
         Schedule schedule = serviceDto.toEntity(workspace, staff);
         scheduleRepository.save(schedule);
 
-        addEventToExternalCalendar(schedule); //외부 api
+        return schedule;
     }
 
     @Transactional
-    public void cancel(Long scheduleId) {
+    public Schedule cancel(Long scheduleId) {
         Schedule schedule = scheduleRepository.findById(scheduleId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.SCHEDULE_NOT_FOUND));
         if (schedule.getStatus() != ScheduleStatus.BEFORE_WORK) throw new BusinessException(ErrorCode.INVALID_SCHEDULE_CANCEL);
         scheduleRepository.deleteById(scheduleId);
+
+        return schedule;
     }
 
-    private void addEventToExternalCalendar(Schedule schedule) {
-        User user = schedule.getStaff().getUser();
-        CalendarPersonalDto calendarPersonalDto = new CalendarPersonalDto(user);
-        CalendarEventDto calendarEventDto = new CalendarEventDto(schedule);
-        externalCalendarService.addEvent(calendarPersonalDto, calendarEventDto);
-    }
 }
